requirements.txt
flask
flask-socketio
python-socketio
eventlet
python-dotenv
livekit-agents==1.2.14
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-cartesia
livekit-api